Create RL_Path_Planner_Project
This file contains the code for the design of the Grid Map, the Markov Decision Process Environment Class and the implementation of PPO with an MLP policy to complete the design of the path planner. It also includes the code used to train, evaluate and test the algorithm with render capability.